+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 Warning: Before deploying in production a bot that can generate heavy traffic, please contact ALE.
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
+## [1.80.0-dotnet.1] - 2020-11-XX
+-   Add HttpManager class to manage a promised queue of request to the server. It allow to have a queue of request and then avoid to much concurrents one. Only 5 concurrents request allowed.
+
+-   Initial use HttpManager with first method HttpService::get .
+
 ## [1.79.2] - 2020-11-18
 -   Same as 1.80.0-dotnet.0